Feat: 여행 단일 조회 기능

+++ src/main/java/com/hifive/yeodam/tour/repository/TourRepository.java
@@ -3,5 +3,5 @@ package com.hifive.yeodam.tour.repository;
 import com.hifive.yeodam.tour.entity.Tour;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TourRepository extends JpaRepository<Tour, Integer> {
+public interface TourRepository extends JpaRepository<Tour, Long> {
 }
